agents/scraping_bee/pyproject.toml
[project]
name = "scraping-bee-agent"
version = "0.1.0"
description = "ADK agent for web scraping using ScrapingBee API"
requires-python = ">=3.9"
dependencies = [
    "google-adk>=0.0.3",
    "requests>=2.25.0",
    "beautifulsoup4>=4.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"